requirements.txt
fastapi
uvicorn[standard]
websockets
google-adk
google-genai
google-cloud-storage
python-dotenv
